⚠ Warning Subpackage (#546)
* refactor: Moved over validation function

* refactor: remove old validate

* refactor: move over process files

* refactor: moved helpers

* refactor: standardiztion of remaining build warning

* refactor; moved warnings into package

* refactor; expose all warnings in package

* refactor: import missing

* refactor: fix import
